Task #4708: switch bf weights and phase calc to double precision, enable tBeamFormer test 3, switch 2nd FFT sz from 2048 to 4096 conform to the bf pipeline design doc. Adapt gpu_math overloads, tiny change in DBP kernel code to adapt.

+++ branches/Cobalt-Task4708-2/RTCP/Cobalt/GPUProc/src/cuda/Kernels/BeamFormerKernel.h
@@ -74,7 +74,7 @@ namespace LOFAR
                              const Parameters &param);
 
       void enqueue(gpu::Stream &queue, PerformanceCounter &counter,
-                   float subbandFrequency, unsigned SAP);
+                   double subbandFrequency, unsigned SAP);
     };
 
     // Specialization of the KernelFactory for
